requirements.txt
flask
flask-cors
flask-socketio
mysql-connector-python
python-dotenv
werkzeug
pillow
gunicorn
